backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
tenacity>=8.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.24.0

# AI/ML & LangChain
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-groq>=0.0.1
sentence-transformers>=2.2.2
openai>=1.0.0
anthropic>=0.18.0
tiktoken>=0.5.0

# Vector DB
pinecone-client>=3.0.0

# Database & Caching
redis>=5.0.0
asyncpg>=0.28.0
sqlalchemy>=2.0.0

# Document Processing
pypdf>=3.17.0
python-docx>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
